component_tests/seed_users/requirements.txt
Faker==37.4.0
psycopg2_binary==2.9.10
